# backend/internal/controller/transport/http/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Emfont Backend API",
    "version": "v1",
    "description": "Emfont font generation API backed by PostgreSQL, MinIO, and HarfBuzz."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/g/{font}": {
      "post": {
        "summary": "Resolve or build font artifacts for a character set",
        "operationId": "generateFont",
        "parameters": [
          {
            "name": "font",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateFontRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing or newly built font artifact URLs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateFontResponse"
                }
              }
            }
          },
          "404": {
            "description": "Font not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or empty JSON request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured limit.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type is not application/json.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit, build admission, or artifact capacity was exceeded.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Build is pending or object storage is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "504": {
            "description": "The font operation timed out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Font build failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/css/{font}": {
      "get": {
        "summary": "Generate a CSS font-face stylesheet",
        "operationId": "getFontCSS",
        "parameters": [
          {
            "name": "font",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "words",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 4096
            }
          },
          {
            "name": "weight",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "min",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Output format for dynamic generation; only applies when words is present.",
            "schema": {
              "type": "string",
              "enum": [
                "woff2"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSS stylesheet.",
            "content": {
              "text/css": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Font not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid query parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit, build admission, or artifact capacity was exceeded.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Build pending or storage unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "504": {
            "description": "The font operation timed out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Font build failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/list": {
      "get": {
        "summary": "List fonts",
        "operationId": "listFonts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Font list.",
            "headers": {
              "X-Next-Cursor": {
                "description": "Cursor for the next page when present.",
                "schema": {
                  "type": "string"
                }
              },
              "Link": {
                "description": "RFC 8288 next-page link when present.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FontListItem"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid cursor, search, or limit.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Font list lookup failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Font service is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "504": {
            "description": "Font list lookup timed out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/info/{fontID}": {
      "get": {
        "summary": "Read font metadata",
        "operationId": "getFontInfo",
        "parameters": [
          {
            "name": "fontID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Font metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FontInfo"
                }
              }
            }
          },
          "404": {
            "description": "Font not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid font identifier.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Font metadata lookup failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Font service is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "504": {
            "description": "Font metadata lookup timed out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_VALUE",
              "INVALID_FORMAT",
              "INVALID_ENUM",
              "REQUIRED",
              "VALUE_TOO_SHORT",
              "VALUE_TOO_LONG",
              "VALUE_TOO_SMALL",
              "VALUE_TOO_LARGE"
            ]
          },
          "message": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "value": {}
        }
      },
      "Problem": {
        "type": "object",
        "required": [
          "code",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "GenerateFontRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "words"
        ],
        "properties": {
          "words": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          },
          "min": {
            "type": "boolean",
            "default": false
          },
          "weight": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ]
          },
          "format": {
            "type": "string",
            "enum": [
              "woff2"
            ]
          }
        }
      },
      "GenerateFontResponse": {
        "type": "object",
        "required": [
          "code",
          "status",
          "message",
          "location",
          "name"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string"
          },
          "location": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "name": {
            "type": "string"
          },
          "weight": {
            "type": "integer"
          },
          "buildMode": {
            "type": "string",
            "enum": [
              "dynamic",
              "static"
            ]
          }
        }
      },
      "FontListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "weight": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "name_zh": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "family": {
            "type": "string"
          },
          "sid": {
            "type": "integer"
          }
        }
      },
      "FontInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object"
          },
          "category": {
            "type": "string"
          },
          "weight": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "tag": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "family": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "sid": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}